+++ dev/checks/__init__.py
@@ -1,0 +1,1 @@
+"""Check modules package."""